move self-written python files to common_python
adopt copy-script / delete unneded file
