+++ common_python/osm_maps_functions.py
@@ -12,11 +12,11 @@ import sys
 import platform
 
 # import custom python packages
-from common_resources import file_directory_functions as fd_fct
-from common_resources import constants
-from common_resources import constants_functions as const_fct
+from common_python import file_directory_functions as fd_fct
+from common_python import constants
+from common_python import constants_functions as const_fct
 
-from common_resources.downloader import Downloader
+from common_python.downloader import Downloader
 
 
 class OsmMaps:
